requirements.txt
Flask==3.0.3
openai-whisper==20240930
ffmpeg-python==0.2.0
tqdm==4.66.4
rich==13.7.1
orjson==3.10.7
